identitymanagementserv/views.py
from django.contrib.auth import authenticate
from rest_framework import viewsets, status
from rest_framework.decorators import action
from rest_framework.response import Response
from rest_framework_simplejwt.tokens import RefreshToken
from .serializers import UserSerializer, LoginSerializer

class AuthViewSet(viewsets.ViewSet):
    """
    A viewset for handling user authentication actions such as registration and login.
    """

    @action(detail=False, methods=['post'])
    def register(self, request):
        """
        Handle user registration.

        Args:
            request (Request): The request object containing user registration data.

        Returns:
            Response: A response object with the serialized user data or error messages.
        """
        serializer = UserSerializer(data=request.data)
        if not serializer.is_valid():
            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
        serializer.save()
        return Response(serializer.data, status=status.HTTP_201_CREATED)

    @action(detail=False, methods=['post'])
    def login(self, request):
        """
        Handle user login.

        Args:
            request (Request): The request object containing login credentials.

        Returns:
            Response: A response object with JWT tokens or error messages.
        """
        serializer = LoginSerializer(data=request.data)
        if not serializer.is_valid():
            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
        user = authenticate(
            username=serializer.validated_data['username'],
            password=serializer.validated_data['password']
        )
        if user is not None:
            refresh = RefreshToken.for_user(user)
            return Response({
                'refresh': str(refresh),
                'access': str(refresh.access_token),
            })
        return Response({'detail': 'Invalid credentials'}, status=status.HTTP_401_UNAUTHORIZED)