pyproject.toml
[tool.poetry]
name = "pricetracker"
version = "0.1.0"
description = ""
authors = ["Gautham P Harish <your-gauthampharish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
channels = "^4.1.0"
redis = "^5.0.7"
djangorestframework-simplejwt = "^5.3.1"
django-rq = "^2.10.2"
django-filter = "^24.2"
django-redis = "^5.4.0"
websocket = "^0.2.1"
websocket-client = "^1.8.0"
celery = "^5.4.0"
websockets = "^12.0"
daphne = "^4.1.2"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
